requirements.txt
dj-database-url==0.4.2
Django==1.10.7
django-salesforce==0.6.9
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2
requests==2.13.0
simplejson==3.10.0
WebOb==1.7.2
whitenoise==3.3.0
wsgi-basic-auth==1.1.0
